LPS-126713 Extract code and removes dependency with journal
